pyproject.toml
[tool.poetry]
name = "image-agent"
version = "0.1.0"
description = "An intelligent AI agent for image processing and analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "image_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
pillow = "^10.0.0"
torch = "^2.0.0"
transformers = "^4.30.0"
fastapi = "^0.100.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
uvicorn = "^0.23.0"
python-multipart = "^0.0.6"
langchain = "^0.0.300"
streamlit = "^1.29.0"
easyocr = "^1.7.1"
google-cloud-vision = "^3.4.4"
google-cloud-translate = "^3.12.0"
opencv-python-headless = "^4.8.1"
openai = "^1.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
mypy = "^1.5.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "B", "Q", "PL", "RUF"]
ignore = []

[tool.ruff.isort]
known-first-party = ["image_agent"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
